Move cocotb.bridge related code to new module, reexport in cocotb.task
