+++ docs/source/newsfragments/3461.change.rst
@@ -1,1 +1,1 @@
-:func:`cocotb.resume` and :func:`cocotb.bridge` are now implemented as decorator functions and not types. All attributes, e.g. ``log``, are no longer available.
+:func:`.resume` and :func:`.bridge` are now implemented as decorator functions and not types. All attributes, e.g. ``log``, are no longer available.
