+++ docs/source/newsfragments/4054.change.rst
@@ -1,1 +1,1 @@
-``cocotb.external`` and ``cocotb.function`` have been renamed to :func:`cocotb.bridge` and :func:`cocotb.resume` to better reflect their intended use case.
+``cocotb.external`` and ``cocotb.function`` have been renamed to :func:`cocotb.task.bridge` and :func:`cocotb.task.resume` to better reflect their intended use case.
