+++ src/cocotb/_scheduler.py
@@ -22,6 +22,7 @@ from typing import Any, Dict, Union
 
 import cocotb
 import cocotb.handle
+from cocotb._bridge import external_state, external_waiter
 from cocotb._exceptions import InternalError
 from cocotb._gpi_triggers import (
     GPITrigger,
@@ -33,91 +34,10 @@ from cocotb._outcomes import Error, Value, capture
 from cocotb._profiling import profiling_context
 from cocotb._py_compat import insertion_ordered_dict
 from cocotb.task import Task, _TaskState
-from cocotb.triggers import Event
 
 # Sadly the Python standard logging module is very slow so it's better not to
 # make any calls by testing a boolean flag first
 _debug = "COCOTB_SCHEDULER_DEBUG" in os.environ
-
-
-class external_state:
-    INIT = 0
-    RUNNING = 1
-    PAUSED = 2
-    EXITED = 3
-
-
-class external_waiter:
-    def __init__(self):
-        self._outcome = None
-        self.thread = None
-        self.event = Event()
-        self.state = external_state.INIT
-        self.cond = threading.Condition()
-        self._log = logging.getLogger(f"cocotb.bridge.{self.thread}.0x{id(self):x}")
-
-    @property
-    def result(self):
-        return self._outcome.get()
-
-    def _propagate_state(self, new_state):
-        with self.cond:
-            if _debug:
-                self._log.debug(
-                    f"Changing state from {self.state} -> {new_state} from {threading.current_thread()}"
-                )
-            self.state = new_state
-            self.cond.notify()
-
-    def thread_done(self):
-        if _debug:
-            self._log.debug(f"Thread finished from {threading.current_thread()}")
-        self._propagate_state(external_state.EXITED)
-
-    def thread_suspend(self):
-        self._propagate_state(external_state.PAUSED)
-
-    def thread_start(self):
-        if self.state > external_state.INIT:
-            return
-
-        if not self.thread.is_alive():
-            self._propagate_state(external_state.RUNNING)
-            self.thread.start()
-
-    def thread_resume(self):
-        self._propagate_state(external_state.RUNNING)
-
-    def thread_wait(self):
-        if _debug:
-            self._log.debug(
-                f"Waiting for the condition lock {threading.current_thread()}"
-            )
-
-        with self.cond:
-            while self.state == external_state.RUNNING:
-                self.cond.wait()
-
-            if _debug:
-                if self.state == external_state.EXITED:
-                    self._log.debug(
-                        f"Thread {self.thread} has exited from {threading.current_thread()}"
-                    )
-                elif self.state == external_state.PAUSED:
-                    self._log.debug(
-                        f"Thread {self.thread} has called yield from {threading.current_thread()}"
-                    )
-                elif self.state == external_state.RUNNING:
-                    self._log.debug(
-                        f"Thread {self.thread} is in RUNNING from {threading.current_thread()}"
-                    )
-
-            if self.state == external_state.INIT:
-                raise Exception(
-                    f"Thread {self.thread} state was not allowed from {threading.current_thread()}"
-                )
-
-        return self.state
 
 
 class Scheduler:
