+++ tests/test_cases/test_async_bridge/test_async_bridge.py
@@ -5,20 +5,22 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 import threading
+import time
 
 import pytest
 
 import cocotb
 from cocotb.clock import Clock
+from cocotb.task import bridge, resume
 from cocotb.triggers import ReadOnly, RisingEdge, Timer
-from cocotb.utils import get_sim_time
+from cocotb.utils import get_sim_steps, get_sim_time
 
 
 def return_two(dut):
     return 2
 
 
-@cocotb.resume
+@resume
 async def await_two_clock_edges(dut):
     await RisingEdge(dut.clk)
     await RisingEdge(dut.clk)
@@ -27,7 +29,7 @@ async def await_two_clock_edges(dut):
     return 2
 
 
-def calls_cocotb_resume(dut):
+def calls_resume(dut):
     return await_two_clock_edges(dut)
 
 
@@ -35,26 +37,22 @@ def print_sim_time(dut, base_time):
     # We are not calling out here so time should not advance
     # And should also remain consistent
     for _ in range(5):
-        _t = get_sim_time("ns")
-        dut._log.info("Time reported = %d", _t)
-        assert _t == base_time
-    dut._log.info("bridge function has ended")
+        time.sleep(0.02)
+        assert get_sim_time("step") == base_time
 
 
 @cocotb.test()
 async def test_time_in_bridge(dut):
     """
     Test that the simulation time does not advance if the wrapped blocking
-    routine does not call @cocotb.resume
+    routine does not call @resume
     """
     await Timer(10, unit="ns")
-    time = get_sim_time("ns")
-    dut._log.info("Time at start of test = %d", time)
-    for i in range(100):
-        dut._log.info("Loop call %d", i)
-        await cocotb.bridge(print_sim_time)(dut, time)
+    time = get_sim_time("step")
+    for i in range(10):
+        await bridge(print_sim_time)(dut, time)
 
-    time_now = get_sim_time("ns")
+    time_now = get_sim_time("step")
     await Timer(10, unit="ns")
 
     assert time == time_now
@@ -63,16 +61,16 @@ async def test_time_in_bridge(dut):
 @cocotb.test()
 async def test_time_in_resume(dut):
     """
-    Test that an @cocotb.bridge function calling back into a cocotb @cocotb.resume
+    Test that an @bridge function calling back into a cocotb @resume
     takes the expected amount of time
     """
 
-    @cocotb.resume
+    @resume
     async def wait_cycles(dut, n):
         for _ in range(n):
             await RisingEdge(dut.clk)
 
-    @cocotb.bridge
+    @bridge
     def wait_cycles_wrapper(dut, n):
         return wait_cycles(dut, n)
 
@@ -81,17 +79,17 @@ async def test_time_in_resume(dut):
     for n in range(5):
         for i in range(20):
             await RisingEdge(dut.clk)
-            time = get_sim_time("ns")
-            expected_after = time + 100 * n
+            time = get_sim_time("step")
+            expected_after = time + get_sim_steps(100, "ns") * n
             await wait_cycles_wrapper(dut, n)
-            time_after = get_sim_time("ns")
+            time_after = get_sim_time("step")
             assert expected_after == time_after
 
 
 @cocotb.test()
 async def test_blocking_function_call_return(dut):
     """
-    Test ability to await a blocking function that is not a coroutine using @cocotb.bridge
+    Test ability to await a blocking function that is not a coroutine using @bridge
     """
 
     async def clock_monitor(dut):
@@ -103,20 +101,20 @@ async def test_blocking_function_call_return(dut):
 
     cocotb.start_soon(clock_monitor(dut))
     cocotb.start_soon(Clock(dut.clk, 100, unit="ns").start())
-    value = await cocotb.bridge(return_two)(dut)
+    value = await bridge(return_two)(dut)
     assert value == 2
 
 
 @cocotb.test()
 async def test_consecutive_bridges(dut):
     """
-    Test that multiple @cocotb.bridge functions can be called in the same test
+    Test that multiple @bridge functions can be called in the same test
     """
-    value = await cocotb.bridge(return_two)(dut)
+    value = await bridge(return_two)(dut)
     dut._log.info("First one completed")
     assert value == 2
 
-    value = await cocotb.bridge(return_two)(dut)
+    value = await bridge(return_two)(dut)
     dut._log.info("Second one completed")
     assert value == 2
 
@@ -124,39 +122,39 @@ async def test_consecutive_bridges(dut):
 @cocotb.test()
 async def test_bridge_from_readonly(dut):
     """
-    Test that @cocotb.bridge functions that don't consume simulation time
+    Test that @bridge functions that don't consume simulation time
     can be called from ReadOnly state
     """
     await ReadOnly()
     dut._log.info("In readonly")
-    value = await cocotb.bridge(return_two)(dut)
+    value = await bridge(return_two)(dut)
     assert value == 2
 
 
 @cocotb.test()
 async def test_resume_from_readonly(dut):
     """
-    Test that @cocotb.bridge functions that call @cocotb.resumes that await Triggers
+    Test that @bridge functions that call @resumes that await Triggers
     can be called from ReadOnly state
     """
     cocotb.start_soon(Clock(dut.clk, 100, unit="ns").start())
 
     await ReadOnly()
     dut._log.info("In readonly")
-    value = await cocotb.bridge(calls_cocotb_resume)(dut)
+    value = await bridge(calls_resume)(dut)
     assert value == 2
 
 
 @cocotb.test()
 async def test_resume_that_awaits(dut):
     """
-    Test that @cocotb.bridge functions can call @cocotb.resume coroutines that
+    Test that @bridge functions can call @resume coroutines that
     awaits Triggers and return values back through to
     the test
     """
     cocotb.start_soon(Clock(dut.clk, 100, unit="ns").start())
 
-    value = await cocotb.bridge(calls_cocotb_resume)(dut)
+    value = await bridge(calls_resume)(dut)
     assert value == 2
 
 
@@ -164,12 +162,12 @@ async def test_resume_that_awaits(dut):
 async def test_await_after_bridge(dut):
     """
     Test that awaiting a Trigger works after returning
-    from @cocotb.bridge functions that call @cocotb.resumes that consume
+    from @bridge functions that call @resumes that consume
     simulation time
     """
     cocotb.start_soon(Clock(dut.clk, 100, unit="ns").start())
 
-    value = await cocotb.bridge(calls_cocotb_resume)(dut)
+    value = await bridge(calls_resume)(dut)
     assert value == 2
 
     await Timer(10, unit="ns")
@@ -179,16 +177,16 @@ async def test_await_after_bridge(dut):
 @cocotb.test()
 async def test_bridge_from_start_soon(dut):
     """
-    Test that @cocotb.bridge functions work when awaited from a forked
+    Test that @bridge functions work when awaited from a forked
     task
     """
 
     async def run_function(dut):
-        value = await cocotb.bridge(calls_cocotb_resume)(dut)
+        value = await bridge(calls_resume)(dut)
         return value
 
     async def run_bridge(dut):
-        value = await cocotb.bridge(return_two)(dut)
+        value = await bridge(return_two)(dut)
         return value
 
     cocotb.start_soon(Clock(dut.clk, 100, unit="ns").start())
@@ -208,10 +206,10 @@ async def test_bridge_from_start_soon(dut):
 @cocotb.test()
 async def test_bridge_raised_exception(dut):
     """
-    Test that exceptions thrown by @cocotb.bridge functions can be caught
+    Test that exceptions thrown by @bridge functions can be caught
     """
 
-    @cocotb.bridge
+    @bridge
     def func():
         raise ValueError()
 
@@ -222,10 +220,10 @@ async def test_bridge_raised_exception(dut):
 @cocotb.test()
 async def test_bridge_returns_exception(dut):
     """
-    Test that exceptions can be returned by @cocotb.bridge functions
+    Test that exceptions can be returned by @bridge functions
     """
 
-    @cocotb.bridge
+    @bridge
     def func():
         return ValueError()
 
@@ -237,14 +235,14 @@ async def test_bridge_returns_exception(dut):
 @cocotb.test()
 async def test_resume_raised_exception(dut):
     """
-    Test that exceptions thrown by @cocotb.resume coroutines can be caught
+    Test that exceptions thrown by @resume coroutines can be caught
     """
 
-    @cocotb.resume
+    @resume
     async def func():
         raise ValueError()
 
-    @cocotb.bridge
+    @bridge
     def ext():
         return func()
 
@@ -255,14 +253,14 @@ async def test_resume_raised_exception(dut):
 @cocotb.test()
 async def test_resume_returns_exception(dut):
     """
-    Test that exceptions can be returned by @cocotb.resume coroutines
+    Test that exceptions can be returned by @resume coroutines
     """
 
-    @cocotb.resume
+    @resume
     async def gen_func():
         return ValueError()
 
-    @cocotb.bridge
+    @bridge
     def ext():
         return gen_func()
 
@@ -274,13 +272,13 @@ async def test_resume_returns_exception(dut):
 @cocotb.test()
 async def test_resume_from_weird_thread_fails(dut):
     """
-    Test that background threads calling a @cocotb.resume do not hang forever
+    Test that background threads calling a @resume do not hang forever
     """
     func_started = False
     caller_resumed = False
     raised = False
 
-    @cocotb.resume
+    @resume
     async def func():
         nonlocal func_started
         func_started = True
@@ -296,7 +294,7 @@ async def test_resume_from_weird_thread_fails(dut):
         finally:
             caller_resumed = True
 
-    @cocotb.bridge
+    @bridge
     def ext():
         t = threading.Thread(target=function_caller)
         t.start()
@@ -316,16 +314,16 @@ async def test_resume_from_weird_thread_fails(dut):
 @cocotb.test()
 async def test_resume_called_in_parallel(dut):
     """
-    Test that the same `@cocotb.resume` can be called from two parallel background
+    Test that the same `@resume` can be called from two parallel background
     threads.
     """
 
-    @cocotb.resume
+    @resume
     async def function(x):
         await Timer(1, unit="ns")
         return x
 
-    @cocotb.bridge
+    @bridge
     def call_function(x):
         return function(x)
